mmu: moving init prior to main
The kernel will be eventually mapped to a high memory region.
This however requires identity mapping since the PC does not
immediately have the newly mapped address. As a result, to make
it convenient, we will map the kernel (from main) fully within
the higher memory range and explicitly jump to it.

Signed-off-by: Utsav Agarwal <utsavagarwal2014@gmail.com>